requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.1
    # via -r requirements.in
aiokafka==0.9.0
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
alembic==1.13.0
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   aiokafka
    #   asyncpg
asyncpg==0.29.0
    # via -r requirements.in
attrs==23.1.0
    # via aiohttp
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via click
exceptiongroup==1.2.0
    # via anyio
fastapi==0.104.1
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.6
    # via
    #   anyio
    #   yarl
mako==1.3.0
    # via alembic
markupsafe==2.1.3
    # via mako
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.2
    # via aiokafka
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.5.2
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.14.5
    # via pydantic
python-dotenv==1.0.0
    # via -r requirements.in
six==1.16.0
    # via transliterate
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.23
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
transliterate==1.10.2
    # via -r requirements.in
typing-extensions==4.8.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
uvicorn==0.24.0.post1
    # via -r requirements.in
yarl==1.9.4
    # via aiohttp
